[Resume] fix detail descriptions & navigate to mypage after download (#361)
* (Fix) navigate to mypage after advanced resume download

* (Feat) pops up download ready modal on resume/upload page

* (Style) change resume button name

* (Style) change resume form subtitle
