+++ src/features/resume/download/ui/Question.tsx
@@ -1,21 +1,18 @@
 import AnswerForm from '@/src/features/resume/download/ui/AnswerForm';
 import { Button } from '@/src/widgets/ui/button';
 import { AILoading } from '@/src/widgets/ui/loader';
-import { StateBasedModal } from '@/src/widgets/ui/modal';
+import AdvancedDownloadModal from './AdvancedDownloadModal';
 
 import {
   useExtraQuestion,
   useAIResponseState,
 } from '@/src/features/resume/download/model/extraQuestionStore';
 import { useAnswer } from '@/src/features/resume/download/model/answerStore';
-import { useDownload } from '@/src/features/resume/download/model/downloadStore';
 import { fetchQuestion } from '@/src/features/resume/download/api/fetchQuestionSecond';
 
-import { useRouter } from 'next/navigation';
 import { useForm } from 'react-hook-form';
 
 export const Question = () => {
-  const router = useRouter();
   const { isLoading } = useAIResponseState();
   const { extraQuestion } = useExtraQuestion();
 
@@ -68,38 +65,7 @@ export const Question = () => {
         </form>
       )}
 
-      <StateBasedModal
-        open={useDownload.getState().isReady}
-        title="이력서 생성에 성공했어요!"
-        description="이력서를 다운로드 받으시겠어요?"
-        actionText="다운로드"
-        cancelText="홈으로"
-        onOpenChange={(open) => {
-          if (!open) {
-            useDownload.getState().setIsReady(false);
-            router.push('/');
-          }
-        }}
-        onAction={() => {
-          const rawUrl = useDownload.getState().url;
-          if (rawUrl) {
-            const link = document.createElement('a');
-            link.href = rawUrl;
-            const defaultFileName = 'resume_download.docx';
-            try {
-              const url = new URL(rawUrl);
-              const pathname = url.pathname;
-              const extractedName = pathname.substring(pathname.lastIndexOf('/') + 1);
-              link.download = extractedName || defaultFileName;
-            } catch {
-              link.download = defaultFileName;
-            }
-            document.body.appendChild(link);
-            link.click();
-            document.body.removeChild(link);
-          }
-        }}
-      />
+      <AdvancedDownloadModal />
     </div>
   );
 };
